requirements.txt
streamlit
pandas
scikit-learn
matplotlib
seaborn
openai
python-docx
